Update literature.txt
The .txt and .doc files are not needed. Docx files cannot be viewed from the repo and the .txt files were just reminders and TODOs for remembering what to do. They contain notes to self and draft comments that should not be visible to collaborators. I kept this file but deleted the informal comments.